+++ settings.py
@@ -106,6 +106,7 @@ GRAPHS_DEFAULT_OPTIONS = {
             facet_col=GraphOption("Subplots", "", clearable=True),
             xaxis_title=GraphOption("X-Axis Title", "", type="input", plotly_option=False),
             yaxis_title=GraphOption("Y-Axis Title", "", type="input", plotly_option=False),
+            subplot_label=GraphOption("Subplot Title", "", type="input", plotly_option=False),
             legend_title=GraphOption("Legend Title", "", type="input", plotly_option=False),
             bargap=GraphOption("Bar Gap", "", type="number", plotly_option=False),
             margin_l=GraphOption("Margin Left", "", type="number", plotly_option=False),